+++ internal/query/options/constants.go
@@ -1,0 +1,8 @@
+package options
+
+type ResultSetsType uint8
+
+const (
+	ResultSetsTypeOrdered = ResultSetsType(iota)
+	ResultSetsTypeConcurrent
+)